requirements.txt
opencv-python-headless
numpy
scipy
kivy
pyjnius
watchdog
